FEATURE - 라인에 존재하는 포인트 목록 객체 생성 - 라인에 있는 포인트 생성 기능 옮기기
