+++ README.md
@@ -12,4 +12,5 @@
 [X] 사다리 포인트 객체 생성
 [X] 사다리 라인 객체 생성
 [X] 한 라인에서 연속으로 연결 안되도록 포인트 생성
+[X] 라인에 존재하는 포인트 목록 객체 생성 - 라인에 있는 포인트 생성 기능 옮기기
 [] 서비스 로직 작업
